% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIACS_batch.R
\name{SIACS.batch}
\alias{SIACS.batch}
\title{Run SIACS Simulations in Batch Mode}
\usage{
SIACS.batch(
  position = 1,
  perturbation = FALSE,
  chemistry = TRUE,
  mechanism = "SAPRC99",
  input_data_list,
  instances,
  OutputList,
  use_parallel = TRUE,
  n_cores = NULL
)
}
\arguments{
\item{position}{Integer, starting position in instances list}

\item{perturbation}{Logical}

\item{chemistry}{Logical}

\item{mechanism}{String, chemical mechanism name}

\item{input_data_list}{List of input data}

\item{instances}{Vector of instance indices}

\item{OutputList}{List of output configurations}

\item{use_parallel}{Logical}

\item{n_cores}{Integer}
}
\description{
Run SIACS Simulations in Batch Mode
}
